Revert channel amount to 200, restore eeprom layout

+++ src/core/misc.c
@@ -192,7 +192,7 @@ volatile bool     gTxTimeoutReached;
     #endif
     #ifdef ENABLE_SERIAL_SCREENCAST
          volatile uint8_t  gUART_LockScreenshot = 0; // lock screenshot if Chirp is used
-         bool gUSB_ScreenshotEnabled = false;
+         bool              gUSB_ScreenshotEnabled = false;
     #endif
 #endif
 
@@ -235,8 +235,7 @@ bool              gFlagReconfigureVfos;
 uint8_t           gVfoConfigureMode;
 bool              gFlagResetVfos;
 bool              gRequestSaveVFO;
-
-uint16_t          gRequestSaveChannel;
+uint8_t           gRequestSaveChannel;
 bool              gRequestSaveSettings;
 #ifdef ENABLE_FMRADIO
     bool          gRequestSaveFM;
@@ -264,7 +263,7 @@ bool              g_SquelchLost;
 volatile uint16_t gFlashLightBlinkCounter;
 
 bool              gFlagEndTransmission;
-uint16_t          gNextMrChannel;
+uint8_t           gNextMrChannel;
 ReceptionMode_t   gRxReceptionMode;
 
 bool              gRxVfoIsActive;
@@ -282,7 +281,7 @@ uint8_t           gFSKWriteIndex;
 
 #ifdef ENABLE_NOAA
     bool          gIsNoaaMode;
-    uint16_t      gNoaaChannel;
+    uint8_t       gNoaaChannel;
 #endif
 
 bool              gUpdateDisplay;
